Set up shortstop to handle env variable in the config
